+++ packages/website/src/components/fluent-layout/fluent-layout.tsx
@@ -1,0 +1,21 @@
+import { useColorMode } from "@docusaurus/theme-common";
+import { FluentProvider, webDarkTheme, webLightTheme } from "@fluentui/react-components";
+import Layout from "@theme/Layout";
+
+export const FluentLayout = ({ children }) => {
+  return (
+    <Layout>
+      <FluentWrapper>{children}</FluentWrapper>
+    </Layout>
+  );
+};
+
+const FluentWrapper = ({ children }) => {
+  const { colorMode } = useColorMode();
+
+  return (
+    <FluentProvider theme={colorMode === "dark" ? webDarkTheme : webLightTheme}>
+      {children}
+    </FluentProvider>
+  );
+};
